requirements.txt
tweepy==4.12.1
python-dotenv==0.19.2